+++ src/RoadRegistry.BackOffice.Abstractions/RoadSegmentsOutline/CreateRoadSegmentOutlineRequest.cs
@@ -4,4 +4,14 @@ using Be.Vlaanderen.Basisregisters.Sqs.Responses;
 using MediatR;
 using Messages;
 
-public sealed record CreateRoadSegmentOutlineRequest(RoadSegmentGeometry Geometry, RoadSegmentStatus Status, RoadSegmentMorphology Morphology, RoadSegmentAccessRestriction AccessRestriction, OrganizationId MaintenanceAuthority, RoadSegmentSurfaceType SurfaceType, RoadSegmentWidth Width, RoadSegmentLaneCount LaneCount, RoadSegmentLaneDirection LaneDirection) : IRequest<ETagResponse>;
+public sealed record CreateRoadSegmentOutlineRequest(
+    RoadSegmentGeometry Geometry,
+    RoadSegmentStatus Status,
+    RoadSegmentMorphology Morphology,
+    RoadSegmentAccessRestriction AccessRestriction,
+    OrganizationId MaintenanceAuthority,
+    RoadSegmentSurfaceType SurfaceType,
+    RoadSegmentWidth Width,
+    RoadSegmentLaneCount LaneCount,
+    RoadSegmentLaneDirection LaneDirection,
+    RoadSegmentCategory Category) : IRequest<ETagResponse>;