+++ src/RoadRegistry.BackOffice.Api/Infrastructure/SchemaFilters/OutlinedRoadSegmentCategorySchemaFilter.cs
@@ -1,0 +1,9 @@
+﻿namespace RoadRegistry.BackOffice.Api.Infrastructure.SchemaFilters;
+
+public class OutlinedRoadSegmentCategorySchemaFilter : EnumSchemaFilter<RoadSegmentCategory.Edit, RoadSegmentCategory>
+{
+    public OutlinedRoadSegmentCategorySchemaFilter()
+        : base(RoadSegmentCategory.Edit.Editable)
+    {
+    }
+}