+++ src/RoadRegistry.BackOffice/FeatureToggles/UseGrbExtractZipArchiveWriterV2FeatureToggle.cs
@@ -1,0 +1,3 @@
+namespace RoadRegistry.BackOffice.FeatureToggles;
+
+public sealed record UseGrbExtractZipArchiveWriterV2FeatureToggle(bool FeatureEnabled) : ApplicationFeatureToggle(FeatureEnabled);