+++ src/RoadRegistry.BackOffice/Messages/RoadNetworkExtractGotRequestedV2.cs
@@ -14,5 +14,6 @@ public class RoadNetworkExtractGotRequestedV2 : IRoadNetworkExtractGotRequestedM
     public string ExternalRequestId { get; set; }
     public string RequestId { get; set; }
     public bool IsInformative { get; set; }
+    public string ZipArchiveWriterVersion { get; set; }
     public string When { get; set; }
 }