+++ src/RoadRegistry.BackOffice/RoadSegmentValidationExtensions.cs
@@ -23,6 +23,11 @@ public static class RoadSegmentValidationExtensions
         return RoadSegmentStatus.Edit.Editable.Contains(status);
     }
 
+    public static bool IsValidForEdit(this RoadSegmentCategory status)
+    {
+        return RoadSegmentCategory.Edit.Editable.Contains(status);
+    }
+
     public static bool IsValidForEdit(this RoadSegmentLaneCount laneCount)
     {
         return laneCount != 0;