+++ test/RoadRegistry.BackOffice.Api.Tests/RoadSegments/WhenCreateOutline/Fixtures/WhenCreateOutlineWithInvalidWegcategorieFixture.cs
@@ -1,0 +1,17 @@
+namespace RoadRegistry.BackOffice.Api.Tests.RoadSegments.WhenCreateOutline.Fixtures;
+
+using Api.RoadSegments;
+using Editor.Schema;
+using MediatR;
+
+public class WhenCreateOutlineWithInvalidWegcategorieFixture : WhenCreateOutlineWithValidRequestFixture
+{
+    public WhenCreateOutlineWithInvalidWegcategorieFixture(IMediator mediator, EditorContext editorContext) : base(mediator, editorContext)
+    {
+    }
+
+    protected override PostRoadSegmentOutlineParameters CreateRequest()
+    {
+        return base.CreateRequest() with { Wegcategorie = RoadSegmentCategory.MainRoad.ToDutchString() };
+    }
+}