+++ test/RoadRegistry.Integration.ProjectionHost.Tests/Projections/Version/RoadSegmentVersionProjectionTests-PartOfEuropeanRoads.cs
@@ -15,7 +15,7 @@ public partial class RoadSegmentVersionProjectionTests
 
         var importedRoadSegment = _fixture.Create<ImportedRoadSegment>();
         importedRoadSegment.PartOfEuropeanRoads = [];
-        
+
         var expectedRecords = Array.Empty<object>()
             .Concat(new[]
             {
@@ -77,6 +77,89 @@ public partial class RoadSegmentVersionProjectionTests
         return BuildProjection()
             .Scenario()
             .Given(acceptedRoadSegmentAdded, message)
+            .Expect(expectedRecords);
+    }
+
+    [Fact]
+    public Task When_adding_road_segment_to_european_road_multiple_times_then_existing_record_is_reused()
+    {
+        _fixture.Freeze<RoadSegmentId>();
+
+        var acceptedRoadSegmentAdded = _fixture
+            .Create<RoadNetworkChangesAccepted>()
+            .WithAcceptedChanges(_fixture.Create<RoadSegmentAdded>());
+
+        var message = _fixture
+            .Create<RoadNetworkChangesAccepted>()
+            .WithAcceptedChanges(_fixture.Create<RoadSegmentAddedToEuropeanRoad>());
+
+        var expectedRecords = BuildInitialExpectedRoadSegmentRecords(acceptedRoadSegmentAdded)
+            .Concat(Array.ConvertAll(message.Changes, change =>
+            {
+                var roadSegmentAddedToEuropeanRoad = change.RoadSegmentAddedToEuropeanRoad;
+
+                var position = 1;
+
+                return BuildRoadSegmentRecord(
+                    position,
+                    acceptedRoadSegmentAdded,
+                    acceptedRoadSegmentAdded.Changes.Single().RoadSegmentAdded,
+                    roadSegment =>
+                    {
+                        roadSegment.Version = roadSegmentAddedToEuropeanRoad.SegmentVersion!.Value;
+
+                        roadSegment.OrganizationId = message.OrganizationId;
+                        roadSegment.OrganizationName = message.Organization;
+                        roadSegment.VersionTimestamp = LocalDateTimeTranslator.TranslateFromWhen(message.When);
+
+                        roadSegment.PartOfEuropeanRoads.Add(new RoadSegmentEuropeanRoadAttributeVersion
+                        {
+                            Position = position,
+                            Id = roadSegmentAddedToEuropeanRoad.AttributeId,
+                            RoadSegmentId = roadSegmentAddedToEuropeanRoad.SegmentId,
+                            Number = roadSegmentAddedToEuropeanRoad.Number,
+                            OrganizationId = message.OrganizationId,
+                            OrganizationName = message.Organization,
+                            CreatedOnTimestamp = LocalDateTimeTranslator.TranslateFromWhen(message.When),
+                            VersionTimestamp = LocalDateTimeTranslator.TranslateFromWhen(message.When)
+                        });
+                    });
+            }))
+            .Concat(Array.ConvertAll(message.Changes, change =>
+            {
+                var roadSegmentAddedToEuropeanRoad = change.RoadSegmentAddedToEuropeanRoad;
+
+                var position = 2;
+
+                return BuildRoadSegmentRecord(
+                    position,
+                    acceptedRoadSegmentAdded,
+                    acceptedRoadSegmentAdded.Changes.Single().RoadSegmentAdded,
+                    roadSegment =>
+                    {
+                        roadSegment.Version = roadSegmentAddedToEuropeanRoad.SegmentVersion!.Value;
+
+                        roadSegment.OrganizationId = message.OrganizationId;
+                        roadSegment.OrganizationName = message.Organization;
+                        roadSegment.VersionTimestamp = LocalDateTimeTranslator.TranslateFromWhen(message.When);
+
+                        roadSegment.PartOfEuropeanRoads.Add(new RoadSegmentEuropeanRoadAttributeVersion
+                        {
+                            Position = position,
+                            Id = roadSegmentAddedToEuropeanRoad.AttributeId,
+                            RoadSegmentId = roadSegmentAddedToEuropeanRoad.SegmentId,
+                            Number = roadSegmentAddedToEuropeanRoad.Number,
+                            OrganizationId = message.OrganizationId,
+                            OrganizationName = message.Organization,
+                            CreatedOnTimestamp = LocalDateTimeTranslator.TranslateFromWhen(message.When),
+                            VersionTimestamp = LocalDateTimeTranslator.TranslateFromWhen(message.When)
+                        });
+                    });
+            }));
+
+        return BuildProjection()
+            .Scenario()
+            .Given(acceptedRoadSegmentAdded, message, message)
             .Expect(expectedRecords);
     }
 
@@ -151,14 +234,14 @@ public partial class RoadSegmentVersionProjectionTests
         _fixture.Freeze<RoadSegmentId>();
 
         var importedRoadSegment = _fixture.Create<ImportedRoadSegment>();
-        
+
         var message = _fixture
             .Create<RoadNetworkChangesAccepted>()
             .WithAcceptedChanges(_fixture.CreateWith<RoadSegmentRemovedFromEuropeanRoad>(x =>
             {
                 x.AttributeId = importedRoadSegment.PartOfEuropeanRoads.First().AttributeId;
             }));
-        
+
         var expectedRecords = Array.Empty<object>()
             .Concat(new[]
             {